// frontend/src/api/models/HourlyBucket.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overmind Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HourlyBucket
 */
export interface HourlyBucket {
    /**
     * 
     * @type {string}
     * @memberof HourlyBucket
     */
    hour: string;
    /**
     * 
     * @type {number}
     * @memberof HourlyBucket
     */
    avgScore?: number | null;
    /**
     * 
     * @type {number}
     * @memberof HourlyBucket
     */
    spanCount?: number;
    /**
     * 
     * @type {number}
     * @memberof HourlyBucket
     */
    avgLatencyMs?: number | null;
    /**
     * 
     * @type {number}
     * @memberof HourlyBucket
     */
    estimatedCost?: number;
}

/**
 * Check if a given object implements the HourlyBucket interface.
 */
export function instanceOfHourlyBucket(value: object): value is HourlyBucket {
    if (!('hour' in value) || value['hour'] === undefined) return false;
    return true;
}

export function HourlyBucketFromJSON(json: any): HourlyBucket {
    return HourlyBucketFromJSONTyped(json, false);
}

export function HourlyBucketFromJSONTyped(json: any, ignoreDiscriminator: boolean): HourlyBucket {
    if (json == null) {
        return json;
    }
    return {
        
        'hour': json['hour'],
        'avgScore': json['avg_score'] == null ? undefined : json['avg_score'],
        'spanCount': json['span_count'] == null ? undefined : json['span_count'],
        'avgLatencyMs': json['avg_latency_ms'] == null ? undefined : json['avg_latency_ms'],
        'estimatedCost': json['estimated_cost'] == null ? undefined : json['estimated_cost'],
    };
}

export function HourlyBucketToJSON(json: any): HourlyBucket {
    return HourlyBucketToJSONTyped(json, false);
}

export function HourlyBucketToJSONTyped(value?: HourlyBucket | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'hour': value['hour'],
        'avg_score': value['avgScore'],
        'span_count': value['spanCount'],
        'avg_latency_ms': value['avgLatencyMs'],
        'estimated_cost': value['estimatedCost'],
    };
}